// lib/types/api.ts
export interface ISBNDBBookResponse {
  title: string
  title_long?: string
  isbn: string
  isbn13: string
  dewey_decimal?: string
  binding: string
  publisher: string
  language: string
  date_published: string
  edition?: string
  pages?: number
  dimensions?: string
  overview?: string
  excerpt?: string
  synopsis?: string
  authors: string[]
  subjects: string[]
  reviews?: Record<string, unknown>[]
  prices?: Record<string, unknown>[]
  related?: Record<string, unknown>[]
}